@@ -10,36 +10,37 @@
     "test:e2e": "vue-cli-service test:e2e"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/fontawesome-svg-core": "^1.3.0",
     "@fortawesome/free-brands-svg-icons": "^5.13.0",
-    "@fortawesome/free-solid-svg-icons": "^5.13.0",
-    "@fortawesome/vue-fontawesome": "^0.1.9",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/vue-fontawesome": "^3.0.0-5",
     "axios": "^0.19.2",
-    "bulma": "^0.8.2",
+    "bulma": "^0.9.4",
     "cross-env": "^7.0.3",
     "leaflet": "1.7.1",
     "leaflet.markercluster": "^1.4.1",
     "localforage": "^1.7.3",
     "register-service-worker": "^1.7.1",
-    "vue": "^2.6.11",
-    "vue-izitoast": "^1.2.1",
-    "vue-meta": "^2.3.3",
-    "vue-router": "^3.1.6",
-    "vuex": "^3.2.0"
+    "vue": "^3.2.45",
+    "@vue/compat": "^3.2.45",
+    "izitoast": "^1.4.0",
+    "vue-meta": "^3.0.0-alpha.9",
+    "vue-router": "^4.1.6",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@babel/core": "7.7.4",
-    "@vue/cli-plugin-e2e-nightwatch": "^4.3.1",
-    "@vue/cli-plugin-pwa": "^4.3.1",
-    "@vue/cli-plugin-unit-mocha": "^4.3.1",
-    "@vue/cli-service": "^4.3.1",
-    "@vue/test-utils": "1.0.0-beta.29",
+    "@vue/cli-plugin-e2e-nightwatch": "^5.0.8",
+    "@vue/cli-plugin-pwa": "^5.0.8",
+    "@vue/cli-plugin-unit-mocha": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "@vue/test-utils": "^2.2.6",
     "chai": "^4.2.0",
     "moxios": "^0.4.0",
     "sass": "^1.57.1",
     "sass-loader": "^8.0.2",
     "sinon": "^7.5.0",
-    "vue-template-compiler": "^2.6.11"
+    "@vue/compiler-sfc": "^3.2.45"
   },
   "browserslist": [
     "> 1%",
