@@ -1,20 +1,23 @@
-import Vue from 'vue';
-import App from './App.vue';
-import Router from "vue-router";
+import {createApp} from 'vue';
+import Root from './App.vue';
 import router from './router';
 import store from './store';
 import './registerServiceWorker';
+import VueIzitoast from './Classes/VueIzitoast';
+import {createMetaManager, plugin as vueMetaPlugin} from 'vue-meta'
+import {FontAwesomeIcon} from '@fortawesome/vue-fontawesome';
+import {installer as HttpService} from '@/Services/HttpService';
 import './libraries';
 
-Vue.use(Router);
+const app = createApp(Root)
+	.use(router)
+	.use(createMetaManager())
+	.use(vueMetaPlugin)
+	.use(store)
+	.use(VueIzitoast)
+	.use(HttpService)
+	.component('FontAwesomeIcon', FontAwesomeIcon);
 
-new Vue({
-	router,
-	store,
-	beforeCreate() {
-		this.$store.dispatch('Uploads/init');
-		this.$store.dispatch('initSettings');
-	},
-	render: h => h(App)
-}).$mount('#app');
+app.mount('#app');
+
 
