@@ -15,7 +15,7 @@
 			<span class="story__name" v-text="story.name"/>
 		</div>
 
-		<StoryEditModal v-if="edit" :active.sync="createModal" @saved="chooseStory"/>
+		<StoryEditModal v-if="edit" v-model:active="createModal" @saved="chooseStory"/>
 	</div>
 </template>
 
