@@ -1,5 +1,3 @@
-import Vue from 'vue';
-import http, {installer as HttpService} from '@/Services/HttpService';
 import {library} from '@fortawesome/fontawesome-svg-core';
 import {
 	faUpload,
@@ -28,29 +26,16 @@ import {
 	faUndo
 } from '@fortawesome/free-solid-svg-icons';
 import {faFacebook} from '@fortawesome/free-brands-svg-icons';
-import Meta from 'vue-meta';
-import {FontAwesomeIcon} from '@fortawesome/vue-fontawesome';
-import VueIziToast from 'vue-izitoast';
 import errorLogger from '@/Services/ErrorLogger';
 
 library.add(faUndo, faUpload, faTimesCircle, faFileImage, faCameraRetro, faCopy, faSignOutAlt, faGlobe, faFacebook, faSearch, faUsers, faAddressCard, faList, faMapMarked, faGlobeAsia, faUserAlt, faChevronDown, faSignInAlt, faExternalLinkSquareAlt, faSearchLocation, faShareAlt, faImages, faSlidersH, faRss, faEdit);
 
-Vue.use(Meta);
-Vue.component('FontAwesomeIcon', FontAwesomeIcon);
-Vue.use(HttpService);
-Vue.use(VueIziToast, {
-	position: 'bottomCenter',
-	timeout: 2500
-});
-Vue.config.productionTip = false;
-Vue.prototype.$bus = new Vue();
-
 if (process.env.NODE_ENV === 'production') {
-	window.onerror = (message, source, lineno, colno, error) => {
-		errorLogger.handle(error);
-	};
-
-	Vue.config.errorHandler = (error, vm) => {
-		errorLogger.handle(error, vm);
-	};
+	// window.onerror = (message, source, lineno, colno, error) => {
+	// 	errorLogger.handle(error);
+	// };
+	//
+	// Vue.config.errorHandler = (error, vm) => {
+	// 	errorLogger.handle(error, vm);
+	// };
 }
